Split out project test files
